requirements.txt
Pyinstaller
python-docx
pandas
numpy
openpyxl